# application/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: JobsHunter API
  version: "1.0.0"
  description: |
    HTTP API for user authentication, job search, profile management, CV upload, and admin email tools.
servers:
  - url: https://ronin-archesporial-lonnie.ngrok-free.dev
    description: Prod
  - url: http://localhost:8081
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Job:
      type: object
      properties:
        score:
          type: integer
          format: int32
        url:
          type: string
          format: uri
      required: [score, url]
    JobHuntResponse:
      type: object
      properties:
        jobsFound:
          type: array
          items:
            $ref: '#/components/schemas/Job'
      required: [jobsFound]
...
    EmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        subject:
          type: string
          nullable: true
          maxLength: 255
        message:
          type: string
        file:
          type: string
          format: binary
          nullable: true
      required: [email, message]
...
    SearchWithSerpRequest:
      type: object
      properties:
        q:
          type: string
          maxLength: 2000
          description: SERP query
        google_domain:
          type: string
          maxLength: 255
          nullable: true
          description: Google domain (e.g., google.ro). See https://serpapi.com/google-domains
        date_posted:
          type: string
          maxLength: 128
          nullable: true
          description: Date filter (today, last_3_days, last_7_days, last_14_days, last_30_days)
        gl:
          type: string
          maxLength: 2
          nullable: true
          description: Country code. See https://serpapi.com/google-countries
        location:
          type: string
          nullable: true
        workType:
          type: string
          enum: [REMOTE, ONSITE, HYBRID]
          nullable: true
        radius:
          type: integer
          format: int32
          nullable: true
          description: Radius in km
        hl:
          type: string
          maxLength: 2
          nullable: true
          description: Language code. See https://serpapi.com/google-languages
      required: [q]
...
paths:
...
  /api/test/email/send:
    post:
      summary: Send custom email (test role only)
      tags: [Test]
      description: Requires ROLE_TEST.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Email sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not test role).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/test/email/sendUsingTemplate:
    post:
      summary: Send templated job email (test role only)
      tags: [Test]
      description: Requires ROLE_TEST.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobHuntResponse'
      responses:
        '200':
          description: Email sent using template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not test role).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/test/searchWithSerp:
    post:
      summary: Search jobs using SerpAPI (test role only)
      tags: [Test]
      description: Requires ROLE_TEST. Searches for jobs using SerpAPI. Returns list of Job objects.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchWithSerpRequest'
      responses:
        '200':
          description: Jobs found successfully. Returns array of Job objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized - authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - requires ROLE_TEST.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/test/testGptModels:
    get:
      summary: Test GPT models (admin only)
      tags: [Test]
      description: Requires ROLE_ADMIN. Tests GPT API connection and models.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: GPT API response
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - requires ROLE_ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...